add tile gid getter and setter

+++ Classes/CommonDefinition.h
@@ -33,7 +33,7 @@ struct GridCoordinate
 	int y;
 };
 
-static const size_t TILE_SIZE = 16;
+static const int TILE_SIZE = 16;
 
 static const char* DEFUALT_FONT = "fonts/Marker Felt.ttf";
